+++ test/integration/test.js
@@ -249,9 +249,26 @@ describe('Declarative Onboarding Integration Test Suite', function performIntegr
                     body.controls = {
                         dryRun: true
                     };
-
-                    originalDns = JSON.parse(JSON.stringify(body.Common.myDns));
-                    body.Common.myDns.nameServers = [];
+                })
+                .then(() => {
+                    const url = `${common.hostname(bigIpAddress, constants.PORT)}${constants.ICONTROL_API}/tm/sys/dns`;
+                    const options = common.buildBody(url, null, bigIpAuth, 'GET');
+                    const retryOptions = {
+                        trials: 3,
+                        timeInterval: 1000
+                    };
+                    return common.sendRequest(options, retryOptions);
+                })
+                .then((response) => {
+                    const dns = JSON.parse(response.body);
+                    originalDns = {
+                        nameServers: dns.nameServers,
+                        search: dns.search
+                    };
+                    body.Common.myDns.nameServers = ['10.10.10.10'];
+                    body.Common.myDns.search = ['example.com'];
+                    assert.notDeepStrictEqual(body.Common.myDns.nameServers, originalDns.nameServers, 'Original and updated nameServers should not match');
+                    assert.notDeepStrictEqual(body.Common.myDns.search, originalDns.search, 'Original and updated search should not match');
                 })
                 .then(() => common.testRequest(
                     body,
@@ -617,6 +634,7 @@ describe('Declarative Onboarding Integration Test Suite', function performIntegr
         let bigIpAuth;
         let oldAuditLink;
         let newAuditLink;
+        let currentState;
 
         before(() => {
             thisMachine = machines[2];
@@ -641,6 +659,9 @@ describe('Declarative Onboarding Integration Test Suite', function performIntegr
                 .then((body) => common.testRequest(body, `${common.hostname(bigIpAddress, constants.PORT)}`
                     + `${constants.DO_API}`, bigIpAuth, constants.HTTP_ACCEPTED, 'POST'))
                 .then(() => common.testGetStatus(20, 60 * 1000, bigIpAddress, bigIpAuth, constants.HTTP_SUCCESS))
+                .then((response) => {
+                    currentState = response.currentConfig.Common;
+                })
                 .catch((error) => logError(error, bigIpAddress, bigIpAuth));
         });
 
@@ -657,6 +678,10 @@ describe('Declarative Onboarding Integration Test Suite', function performIntegr
                     assert.ok(auditLink);
                 })
                 .catch((error) => logError(error, bigIpAddress, bigIpAuth));
+        });
+
+        it('should have hostname set', () => {
+            assert.deepStrictEqual(currentState.System.hostname, 'do-integration-test.local');
         });
 
         it('should have re-licensed with new pool', () => {
@@ -690,7 +715,6 @@ describe('Declarative Onboarding Integration Test Suite', function performIntegr
         describe('Test Firewall', function testFirewall() {
             this.timeout(1000 * 60 * 30); // 30 minutes
             let body;
-            let currentState;
 
             before(() => {
                 const bodyFile = `${BODIES}/firewall.json`;
@@ -845,7 +869,6 @@ describe('Declarative Onboarding Integration Test Suite', function performIntegr
         });
 
         describe('Test GSLB', function testGslb() {
-            let currentState;
             let body;
 
             before(() => {