+++ test/unit/lib/doUtilTests.js
@@ -455,14 +455,18 @@ describe('doUtil', () => {
                 });
         });
 
-        it('should wait for extra services to be running if specified', () => {
-            bigIpListSpy = sinon.stub(bigIpMock, 'list').resolves(
-                {
+        it('should wait for extra services to be running or down due to provisioning if specified', () => {
+            bigIpListSpy = sinon.stub(bigIpMock, 'list').callsFake((path) => {
+                const response = {
                     apiRawValues: {
                         apiAnonymous: 'run'
                     }
+                };
+                if (path === '/tm/sys/service/myExtraService2/stats') {
+                    response.apiRawValues.apiAnonymous = 'myExtraService2 down, Not provisioned\n';
                 }
-            );
+                return Promise.resolve(response);
+            });
 
             return doUtil.restartService(bigIpMock, 'myService', ['myExtraService1', 'myExtraService2'])
                 .then(() => {
@@ -506,6 +510,24 @@ describe('doUtil', () => {
             return doUtil.restartService(bigIpMock, 'myService')
                 .then(() => {
                     assert.strictEqual(bigIpListSpy.callCount, 1);
+                });
+        });
+
+        it('should wait for child process to complete in case of dhclient', () => {
+            bigIpListSpy = sinon.stub(bigIpMock, 'list').callsFake(() => {
+                const response = {
+                    apiRawValues: {
+                        apiAnonymous: 'run'
+                    }
+                };
+                return Promise.resolve(response);
+            });
+
+            const doUtilStub = sinon.stub(doUtil, 'executeBashCommandIControl').resolves();
+            return doUtil.restartService(bigIpMock, 'dhclient')
+                .then(() => {
+                    assert.strictEqual(doUtilStub.callCount, 1);
+                    assert.strictEqual(bigIpListSpy.args[0][0], '/tm/sys/service/dhclient/stats');
                 });
         });
     });